pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "traffic-signal-env"
version = "1.0.0"
description = "Real-world traffic signal control environment for OpenEnv RL agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["openenv", "reinforcement-learning", "traffic", "urban-mobility"]
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.3",
    "requests>=2.32.3",
    "openai>=1.57.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.openenv]
name = "traffic-signal-control"
version = "1.0.0"
entry_point = "server.app:app"
tasks = [
    "single_intersection_easy",
    "arterial_corridor_medium",
    "urban_grid_hard",
]
